+++ front/app/containers/Admin/projects/project/phaseSetup/utils.ts
@@ -1,3 +1,6 @@
+import { v4 as uuidv4 } from 'uuid';
+
+import { IFileAttachmentData } from 'api/file_attachments/types';
 import { IPhaseData } from 'api/phases/types';
 
 export function getTimelineTab(
@@ -28,3 +31,39 @@ export function getTimelineTab(
 
   return 'setup';
 }
+
+type GenerateTemporaryFileAttachmentParams = {
+  fileId: string;
+  phaseId: string | undefined;
+  position: number;
+};
+
+export const generateTemporaryFileAttachment = ({
+  fileId,
+  phaseId,
+  position,
+}: GenerateTemporaryFileAttachmentParams): IFileAttachmentData => {
+  const temporaryFileAttachment: IFileAttachmentData = {
+    id: `TEMP-${uuidv4()}`, // Temporary ID, to mark it as a newly added file attachment.
+    attributes: {
+      position,
+    },
+    relationships: {
+      attachable: {
+        data: {
+          type: 'Phase',
+          id: phaseId || '',
+        },
+      },
+      file: {
+        data: {
+          id: fileId,
+          type: 'file',
+        },
+      },
+    },
+    type: 'file_attachment',
+  };
+
+  return temporaryFileAttachment;
+};
